requirements.txt
attrs==17.4.0
enum==0.4.6
watchdog==0.8.3

